geoloc/requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
faiss-cpu>=1.7
torch>=2.0
torchvision>=0.15
pandas>=2.0
geopy>=2.4
